feat: Refactor vector creation and editing to utilize toValidFeatureCollection utility
